+++ ChangeLog.md
@@ -1,8 +1,13 @@
 # Revision history for Z-Data
 
+## 0.1.3.0  -- 2020-09-19
+
+* Add indexing funtion to `Z.Data.Vector` and `Z.Data.Text`.
+* Rename `allocMutableByteArrayUnsafe` to `allocBytesUnsafe`, add `allocBytesSafe`.
+
 ## 0.1.2.0  -- 2020-09-19
 
-* Rename 'read/write/indexWord8ArrayAs' to 'read/write/indexWord8ArrayAs#'.
-* Add 'read/write/indexWord8ArrayAs', 'read/write/indexPrimWord8ArrayAs'.
+* Rename `read/write/indexWord8ArrayAs` to `read/write/indexWord8ArrayAs#`.
+* Add `read/write/indexWord8ArrayAs`, `read/write/indexPrimWord8ArrayAs`.
 * Fix JSON encoding code in generic instance(constructor with single payload case).